+++ database/models/Plan.rb
@@ -26,7 +26,7 @@ class Plan < Sequel::Model
       return plan
     end
     plan
-  end
+    end
 
   def self.plan_id_validation(plan_id)
     if plan_id.nil?
@@ -37,34 +37,71 @@ class Plan < Sequel::Model
     []
   end
 
-  # @param [Hash] data
-  # example: {'plan_data': {'product_id': id, 'name': name}} or {'plan_data': {'product_name': name, 'name': name}}
-  def self.create_new(data)
-    other_data = {} # product_id
-    product = if data['plan_data']['product_id'].nil?
-                Product.find_or_create(name: data['plan_data']['product_name'])
-              else
-                Product[id: data['plan_data']['product_id']]
-              end
-    other_data[:product_id] = product.id
-    begin
-      plan = Plan.find_or_create(name: data['plan_data']['name'], product_id: other_data[:product_id]) do |plan|
-        plan.name = data['plan_data']['name']
+  def self.getting_product(data)
+    if data['plan_data']['product_id'].nil?
+      new_product = Product.new(name: data['plan_data']['product_name'])
+      new_product.valid?
+      if new_product.valid?
+        Product.find_or_create(name: data['plan_data']['product_name'])
+      else
+        new_product
       end
-    rescue StandardError
-      return product_id_validation(Plan.new(data['plan_data']), other_data[:product_id])
+    else
+      Product[id: data['plan_data']['product_id']]
     end
-    product.add_plan(plan)
+  end
+
+  def self.find_or_new(data, product_id)
+    Plan.find(name: data, product_id: product_id) || Plan.new(name: data)
+  end
+
+  # @param [Hash] data
+  # example: data = {'plan_data': {'product_id': id, 'name': name}} or {'plan_data': {'product_name': name, 'name': name}}
+  # return responce = {product: {product_data}, plan: {plan_data}, errors:: {product_errors: {}, plan_errors: {}}}
+  def self.create_new(data)
+    return { plan: Plan[id: data['run_data']['plan_id']] } if plan_id_exist?(data)
+    product_resp = create_product(data)
+    if product_resp[:product].nil?
+      { plan_errors: 'product creating error' }.merge(product_resp)
+    else
+      new_plan = Plan.find_or_new(data['plan_data']['name'], product_resp[:product].id)
+      if new_plan.valid?
+        new_plan.save
+        product_resp[:product].add_plan(new_plan)
+        { plan: new_plan }.merge(product_resp)
+      else
+        { plan_errors: new_plan.errors.full_messages }
+      end
+    end
+  end
+
+  def self.create_product(data)
+    unless data['plan_data'].nil?
+      unless data['plan_data']['product_id'].nil? && data['plan_data']['product_name'].nil?
+        return Product.create_new(data['plan_data']['product_id'] || data['plan_data']['product_name'])
+      end
+    end
+    return { product_errors: 'product id of name not found' }
+  end
+
+  def self.plan_id_exist?(data)
+    return !data['run_data']['plan_id'].nil? unless data['run_data'].nil?
+    false
   end
 
   # @param data [Hash] like {'plan_data' => {id: int, plan_name: str}}
   def self.edit(data)
     plan = Plan[id: data['plan_data']['id']]
-    plan.update(name: data['plan_data']['plan_name'], updated_at: Time.now)
-    plan.valid?
-    { 'plan_data' => plan.values, 'errors' => plan.errors }
-  rescue StandardError
-    { 'plan_data' => Plan.new.values, 'errors' => {params: 'Plan data is incorrect FIXME!!'} } # FIXME: add validate
+    if plan.set(name: data['plan_data']['plan_name']).valid?
+      if Plan[name: data['plan_data']['plan_name'], product_id: plan.product_id].nil?
+        plan.update(name: data['plan_data']['plan_name'], updated_at: Time.now)
+        { plan: plan.values }
+      else
+        { plan_errors: ['plan name already used'] }
+      end
+    else
+      { plan_errors: plan.errors.full_messages }
+    end
   end
 
   def self.get_runs(*args)