+++ Changelog.md
@@ -1,6 +1,9 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 ## [0.0.2] - 2018-02-1
+### Added
+    - Add errors after creating product without name
+    - Test for product validation
 ### Fixed
     - Fix errors after name change
 ## [0.0.1] - 2018-01-30